Added a script that will override cypress script
The cypress software tools have a cool feature. When a device that has
an appropriate IDs is plugged in the script at cy_renumerate.sh is
called. This allows their code to be notified using signals. This script
replaces their script with one that will also call prometheus_comm as
well.